requirements.txt
appdirs==1.4.0
dj-database-url==0.4.2
Django==1.10
django-mathfilters==0.4.0
django-widget-tweaks==1.4.1
gunicorn==19.7.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.7.1
pyparsing==2.1.10
six==1.10.0
whitenoise==3.3.0
